switch uncertainty to float32

+++ scripts/create_test_data.py
@@ -41,7 +41,7 @@ def create_header_validation_test_data(path="../punchbowl/tests/"):
     h = m.to_fits_header()
 
     data = np.ones((2, 4096, 4096), dtype=np.float32)
-    uncertainty = StdDevUncertainty(np.sqrt(np.abs(data)).astype(np.float64))
+    uncertainty = StdDevUncertainty(np.sqrt(np.abs(data)).astype(np.float32))
 
     d = PUNCHData(data=data, uncertainty=uncertainty, wcs=WCS(h), meta=m)
     file_path = os.path.join(path, "test_header_validation.fits")
